add eoapi.auth
