+++ dockerfiles/Dockerfile.raster
@@ -6,9 +6,10 @@ ENV CURL_CA_BUNDLE /etc/ssl/certs/ca-certificates.crt
 
 RUN python -m pip install psycopg[binary,pool]
 
+COPY runtimes/eoapi/auth /tmp/auth
 COPY runtimes/eoapi/raster /tmp/raster
-RUN python -m pip install /tmp/raster
-RUN rm -rf /tmp/raster
+RUN python -m pip install /tmp/auth /tmp/raster
+RUN rm -rf /tmp/raster /tmp/auth
 
 ENV MODULE_NAME eoapi.raster.app
 ENV VARIABLE_NAME app
