+++ runtimes/eoapi/auth/eoapi/auth/__init__.py
@@ -1,0 +1,6 @@
+"""eoapi.auth."""
+
+from .auth import OidcAuth, Scope  # noqa
+from .config import AuthSettings  # noqa
+
+__version__ = "0.1.0"
