+++ runtimes/eoapi/auth/eoapi/auth/errors.py
@@ -1,0 +1,5 @@
+"""eoapi.auth errors."""
+
+
+class OidcFetchError(Exception):
+    pass
