+++ runtimes/eoapi/raster/pyproject.toml
@@ -20,12 +20,11 @@ classifiers = [
 ]
 dynamic = ["version"]
 dependencies = [
+    "eoapi.auth",
     "titiler.pgstac==1.3.0",
     "titiler.extensions",
     "starlette-cramjam>=0.3,<0.4",
     "importlib_resources>=1.1.0;python_version<'3.9'",
-    "pyjwt",
-    "cryptography",
 ]
 
 [project.optional-dependencies]
